Added sending timestamps to kafka
